Geary.ImapError: Add new OPERATIONAL_ERROR code
Used to indicate a NO response from an IMAP server when there is no
more appropriate error.

Mark it as being a recoverable failure so that the replay queue will
re-try the command which is useful for #502, since a NO is followed
immediately by a BYE, but we don't get any useful indication that the
NO is because the connection has become unavailable.